moar keyboard shortcuts
