+++ src/util.js
@@ -1,7 +1,6 @@
 import GLib from "gi://GLib";
 import Gio from "gi://Gio";
 import Xdp from "gi://Xdp";
-import { rangeEquals } from "./lsp/LSP.js";
 
 export const portal = new Xdp.Portal();
 
@@ -173,27 +172,4 @@ export function unstack(fn) {
       latest_promise = fn(...latest_arguments);
     });
   };
-}
-
-export function getItersAtRange(buffer, { start, end }) {
-  let start_iter;
-  let end_iter;
-
-  // Apply the tag on the whole line
-  // if diagnostic start and end are equals such as
-  // Blueprint-Error 13:12 to 13:12 Could not determine what kind of syntax is meant here
-  if (rangeEquals(start, end)) {
-    [, start_iter] = buffer.get_iter_at_line(start.line);
-    [, end_iter] = buffer.get_iter_at_line(end.line);
-    end_iter.forward_to_line_end();
-    start_iter.forward_find_char((char) => char !== "", end_iter);
-  } else {
-    [, start_iter] = buffer.get_iter_at_line_offset(
-      start.line,
-      start.character
-    );
-    [, end_iter] = buffer.get_iter_at_line_offset(end.line, end.character);
-  }
-
-  return [start_iter, end_iter];
 }
